Laget ny versjon av gammel løype for beregning av andeler og trukket ut felles kode for ny og gammel løype i egen fil
